docs: add comments and restructure files
